#2 Make query and test for getting available event tickets and max allocation
